+++ doc/source/changelog.rst
@@ -73,6 +73,9 @@ organisation on `GitHub <https://github.com/openbiosim/sire>`__.
   means that we can catch out-of-memory errors and raise a more
   informative exception.
 
+* Fixed the bug where the wrong return type from ``.minimisation()`` and
+  ``.dynamics()`` was returned. This fixes issue #137.
+
 * Please add an item to this changelog when you create your PR
 
 `2023.4.1 <https://github.com/openbiosim/sire/compare/2023.4.0...2023.4.1>`__ - October 2023